refactor: standardize naming conventions and improve code readability in Application and crypto modules by renaming functions and variables, introducing constants for cryptographic sizes, and enhancing error handling in VaultService and PrekeyManager.

+++ src/ui/conversations/ConversationListModel.cpp
@@ -9,7 +9,7 @@ int ConversationListModel::rowCount(const QModelIndex &parent) const {
   if (parent.isValid()) {
     return 0;
   }
-  return m_conversations.size();
+  return static_cast<int>(m_conversations.size());
 }
 
 QVariant ConversationListModel::data(const QModelIndex &index, int role) const {
